Move path resolves into svelte config
