+++ framerail/svelte.config.js
@@ -1,17 +1,37 @@
 import adapter from "@sveltejs/adapter-node"
-import preprocess from "svelte-preprocess"
+import { statSync } from "fs"
+import { dirname, resolve } from "path"
+import { sveltePreprocess } from "svelte-preprocess"
+import { fileURLToPath } from "url"
+
+// The former only works on node 20.11+
+const __dirname = import.meta.dirname ?? dirname(fileURLToPath(import.meta.url))
+
+function resolveAssets() {
+  try {
+    let globalAssets = statSync(resolve(__dirname, "../assets"))
+    if (globalAssets.isDirectory()) return resolve(__dirname, "../assets")
+    else return resolve(__dirname, "src/assets")
+  } catch (error) {
+    return resolve(__dirname, "src/assets")
+  }
+}
 
 /** @type {import("@sveltejs/kit").Config} */
 const config = {
   // Consult https://github.com/sveltejs/svelte-preprocess
   // for more information about preprocessors
-  preprocess: preprocess(),
+  preprocess: sveltePreprocess(),
 
   kit: {
     adapter: adapter(),
     csrf: {
       // Allow flexible hosts on local, since we don't have real DNS
       checkOrigin: process.env.FRAMERAIL_ENV !== "local"
+    },
+    alias: {
+      "$static": resolve(__dirname, "static"),
+      "$assets": resolveAssets()
     }
   },
