+++ framerail/vite.config.ts
@@ -1,17 +1,5 @@
 import { sveltekit } from "@sveltejs/kit/vite"
-import { statSync } from "fs"
-import { resolve } from "path"
 import type { UserConfig } from "vite"
-
-function resolveAssets() {
-  try {
-    let globalAssets = statSync(resolve(__dirname, "../assets"))
-    if (globalAssets.isDirectory()) return resolve(__dirname, "../assets")
-    else return resolve(__dirname, "src/assets")
-  } catch (error) {
-    return resolve(__dirname, "src/assets")
-  }
-}
 
 const config: UserConfig = {
   server: {
@@ -19,12 +7,7 @@ const config: UserConfig = {
     port: 3000,
     strictPort: true
   },
-  plugins: [sveltekit()],
-  resolve: {
-    alias: {
-      "$static": resolve(__dirname, "static"),
-      "$assets": resolveAssets()
-    }
+  plugins: [sveltekit()]
   }
 }
